refactor(!): comply with test suite

+++ src/token.rs
@@ -4,11 +4,11 @@ use std::fmt::Display;
 pub struct Token {
     pub token_type: TokenType,
     pub lexeme: String,
-    pub line: u32, // TODO: Line numbers on tokens
+    pub line: usize,
 }
 
 impl Token {
-    pub fn new(token_type: TokenType, lexeme: String, line: u32) -> Token {
+    pub fn new(token_type: TokenType, lexeme: String, line: usize) -> Token {
         Token {
             token_type,
             lexeme,
